requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
structlog==24.1.0
google-cloud-logging==3.9.0
